+++ webpack.config.js
@@ -15,15 +15,6 @@ Encore
         return `${moduleFileName}`;
     };
 })
-.addExternals({
-    '@hundh/contao-utils-bundle': 'utilsBundle'
-})
-.configureBabel(null,
-{
-    includeNodeModules: [
-        '@hundh/contao-utils-bundle',
-    ]
-})
 .enableSassLoader()
 .enablePostCssLoader()
 .enableSourceMaps(!Encore.isProduction())